pyproject.toml
[project]
name = "scribe"
version = "0.1.0"
description = "A package for single-cell RNA sequencing data analysis using variational inference"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Manuel Razo" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3"
]

# Core dependencies
dependencies = [
    "anndata>=0.11.1",
    "arviz>=0.20.0",
    "graphviz>=0.20.3",
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.3",
    "myst-parser>=4.0.0",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "scanpy>=1.10.4",
    "scipy>=1.14.1",
    "seaborn>=0.13.2",
    "sphinx>=8.1.3",
    "sphinx-rtd-theme>=3.0.2",
    "statsmodels>=0.14.4",
    "numpyro==0.17.0",
    "blackjax==1.2.5",
    "multipledispatch>=1.0.0",
]

# Development dependencies
[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mpl>=0.17.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["scribe*"]
